+++ packages/widgetbook/lib/src/addons/grid_addon/addon.dart
@@ -1,0 +1,2 @@
+export 'grid_addon.dart';
+export 'grid_painter.dart';