+++ packages/widgetbook_cli/bin/cli_staging.dart
@@ -1,7 +1,7 @@
-import 'flavor/flavor.dart';
+import 'package:widgetbook_cli/widgetbook_cli.dart';
+
 import 'main.dart' as cli;
 
 void main(List<String> args) {
-  Flavor().strategy = DeploymentStrategy.staging;
-  cli.main(args);
+  cli.main(args, StagingEnv());
 }