+++ packages/widgetbook_generator/lib/src/tree/tree.dart
@@ -5,19 +5,6 @@ import 'duplicate_use_cases_error.dart';
 import 'tree_node.dart';
 
 class Tree {
-  /// Splits the [uri] into its parts, skipping both the `package:` and
-  /// the `src` parts.
-  ///
-  /// For example, `package:widgetbook/src/widgets/foo/bar.dart`
-  /// will be split into `['widgets', 'foo']`.
-  static List<String> getPathParts(String uri) {
-    final directory = path.dirname(uri);
-    final parts = path.split(directory);
-    final hasSrc = parts.length >= 2 && parts[1] == 'src';
-
-    return parts.skip(hasSrc ? 2 : 1).toList();
-  }
-
   /// Builds a tree from the given [useCases].
   ///
   /// The tree is built by splitting the import uri of each use case into its
@@ -30,7 +17,7 @@ class Tree {
     final root = TreeNode<Null>(null);
 
     for (var useCase in useCases) {
-      final parts = getPathParts(useCase.component.importUri);
+      final parts = path.split(useCase.navPath);
 
       TreeNode currentNode = root;